Upgrade RuboCop to 0.58.1
This version includes a bugfix (https://github.com/rubocop-hq/rubocop/issues/6092)
that should allow us to add PayPal-related code to the RuboCop checks.
For this patch, though, only RuboCop is upgraded (and a tiny tweak to
Mercury implemented to handle a new RuboCop test); the PayPal compliance
changes will be a separate commit.